+++ src/pages/Search.jsx
@@ -4,6 +4,9 @@ import { Link } from "react-router-dom";
 import Card from "../components/Card";
 // import authHeader from "../services/auth-header";
 import api from "../services/api"
+import Loading from "../components/Loading";
+import * as loadingData from "../loading/Search.json"
+import Swal from 'sweetalert2'
 
 // const URL = import.meta.env.VITE_BASE_URL;
 // const USERNAME = import.meta.env.VITE_BASE_USERNAME;
@@ -19,15 +22,17 @@ import api from "../services/api"
 const Search = () => {
   const [data, setData] = useState([]);
   const [searchTerm, setSearchTerm] = useState("");
-
+  const [loading, setLoading] = useState();
 
   const fetchData = async () => {
+    setLoading(true);  // Set loading to true before fetching data
     try {
       const res = await api.get(`/restaurants`);
       setData(res.data);
     } catch (error) {
       console.error("Error fetching data:", error);
     }
+    setLoading(false);  // Set loading to false after fetching data
   };
 
   useEffect(() => {
@@ -35,37 +40,74 @@ const Search = () => {
   }, []);
 
   
+// const handleDelete = async (id) => {
+//   try {
+//     await api.delete(`/restaurants/${id}`);
+//     fetchData();
+//   } catch (error) {
+//     console.error("Error deleting restaurant:", error);
+//   }
+// };
+
+
 const handleDelete = async (id) => {
-  try {
-    await api.delete(`/restaurants/${id}`);
-    fetchData();
-  } catch (error) {
-    console.error("Error deleting restaurant:", error);
-  }
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then( async (result) => {
+    if (result.isConfirmed) {
+      try {
+        await api.delete(`/restaurants/${id}`);
+        await Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+        window.location.reload();
+      } catch (error) {
+        console.error(error);
+      }
+    }
+  })
+  // try {
+  //   await api.delete(`/restaurants/${id}`);
+  //   window.location.reload();
+  // } catch (error) {
+  //   console.error(error);
+  // }
 };
 
 
 return (
-    <div className="search-container">
-      <h1>Restaurant</h1>
-      <input className="search-input"
-        type="text"
-        placeholder="Search..."
-        value={searchTerm}
-        onChange={(e) => setSearchTerm(e.target.value)}
-      />
+  <div className="search-container">
+    <h1>Restaurant</h1>
+    <input className="search-input"
+      type="text"
+      placeholder="Search..."
+      value={searchTerm}
+      onChange={(e) => setSearchTerm(e.target.value)}
+    />
 
-      {data
-        .filter((restaurant) =>
+    {
+      !loading ? (
+        data.filter((restaurant) =>
           restaurant.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
           restaurant.type.toLowerCase().includes(searchTerm.toLowerCase())
         )
-
         .map((filteredRestaurant, index) => (
           <Card key={index} restaurant={filteredRestaurant} handleDelete={handleDelete} />
-        ))}
-    </div>
-  );
+        ))
+      ) : (
+        <Loading animation={{ ...loadingData }} />
+      )
+    }
+  </div>
+);
 }
 
 export default Search;
